Deprecate `get_version_info` and recommend against runtime dependency
