+++ tests/test_deprecated.py
@@ -126,3 +126,11 @@ def test__wrap_command():
     cmd(dist).run()
     assert called == True
 
+
+@fail_if_not_removed
+def test_get_version_info():
+    assert pkg.get_version_info('1.3.0') == (1,3,0)
+    assert pkg.get_version_info('1.3.0a1') == (1,3,0,'a1')
+    assert pkg.get_version_info('1.3.0.dev0') == (1,3,0,'.dev0')
+    with pytest.raises(ValueError):
+        pkg.get_version_info('1.3.0foo1')
